Update Debian packaging files
The Debian packaging files in the tree were ancient.  Update to the
packaging files from the just-uploaded 1.5.73.3-1 version, which should
build and correctly install the current development source.

The current packaging files no longer use a hard-coded version in
debian/rules, so remove the configure machinery for creating it.  The
correct way of handling the version number is to generate a new
changelog entry.  That work will come in a subsequent patch.

Change-Id: I187fe7cc3f4ca38de6165d58274841d300b6cdf1
Reviewed-on: http://gerrit.openafs.org/1713
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>