Remove special-case call sequence for KAM_SetPassword on s390
For Linux s390 (but not s390x), an additional argument was passed
to KAM_SetPassword between the kvno and the encryption key.  This
doesn't seem to match the rest of the code and is now, with stricter
prototyping, preventing the code from compiling.  Remove it and use
the same call sequence on s390 as everywhere else.

Change-Id: I750f64212e80d4bed4500e316ad7d354608866d7
Reviewed-on: http://gerrit.openafs.org/1790
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>
(cherry picked from commit d80382efc5b0e3c8d32688a94c7ff7c68882cb59)