Flesh out NEWS for recent work
Add entries to NEWS for all development releases back to 1.5.36, which
is the point at which I got bored, taken from the release announcements
to openafs-announce.  Try to use a relatively readable and consistent
formatting for all the entries and convert the old entries to the new
format (mostly).

Further history can be added from the release announcements when I or
someone else gets bored.

Change-Id: Iaf9171c82ca74cd98be3745d9c1965efc39d889a
Reviewed-on: http://gerrit.openafs.org/2036
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>