Clean up the introduction to the Administration Guide
Provide a more useful abstract and remove the (outdated) specific list
of supported platforms and the M.m version number placeholders.  Update
the list of associated documents to match their current titles, and
provide a better description of the Reference Manual.

Reformat the parent document and preface for easier maintenance in the
future.

Change-Id: I42ce78274ed7c4ca7a2f0b9c5ec2e6f7a786adb8
Reviewed-on: http://gerrit.openafs.org/2043
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>