Comprehensive edit of chapter one of the Administration Guide
Update and revise chapter one (An Overview of OpenAFS Administration)
of the Administration Guide for current AFS and current computing
concepts.

Replace the Kerberos Server terminology with Kerberos KDC and add
additional details about the relationship between AFS and a Kerberos
KDC.  Remove some remaining Authentication Server references.  Add
some details about the Protection Server management of the mapping
from Kerberos principals to AFS IDs.

Remove some now-obsolete distinctions and concepts between mainframes
and workstations and recommendations for server systems.

Reorganize the order in which the servers are discussed to follow a
somewhat more natural order.

Be clear that the Backup Server is optional and that there are other
methods available to back up AFS.  Mention backing up to disk as well
as tape in a few places.

Change-Id: I57ce083a84ca2a44f7a4383d80b05508e6448284
Reviewed-on: http://gerrit.openafs.org/2045
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>